+++ plutarch-orphanage/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## 1.0.2 -- 27-06-2024
+
+### Added
+
+* Orphan instances for V1 ledger types
+
 ## 1.0.1 -- 24-06-2024
 
 ### Added